[feat] 마이페이지- 알림 Ui 수정
